+++ src/features/categories/FilterableCategoryItems.tsx
@@ -1,16 +1,15 @@
-import { Container } from '@mantine/core';
 import React from 'react';
 
-import SearchInput from '../SearchInput';
+import SearchInput from '../../components/SearchInput';
 import CategoryItems from './CategoryItems';
 
 const FilterableCategoryItems = () => {
   return (
-    <Container className="py-10">
+    <>
       <SearchInput />
       <br />
       <CategoryItems />
-    </Container>
+    </>
   );
 };
